Introduce data_daemon, data_transformer, and data_uploader
Along with the data_puller, they form the components of a lambda-based daemon that will regularly pull data from mock server and upload a snapshot to S3. Some of the business logic are stubs and are not yet implemented. For testing, the main API gateway lambda will temporarily point to the data_daemon handler